Refactor: [002] 공유방 간단한 주석 및 history 삭제, 구조 변경
